Move 3D solver utils to utility3d
