+++ thetis/__init__.py
@@ -1,5 +1,6 @@
 from __future__ import absolute_import
 from thetis.utility import *
+from thetis.utility3d import *
 from thetis.log import *
 import thetis.timeintegrator as timeintegrator  # NOQA
 import thetis.solver as solver  # NOQA
